fix off site viewing
